Extract duplicated util code
